Update Arduino INI file
Final version of code that uses a 20x4 display and shows the set voltage when you need to change with a Potentiometer. Then Displays the final reading on screen for reference. Then also reads the input A1 plasma voltage from a 50:1 divider board as the plasma is on and cutting.